Reuse test skip condition in tearDown()
Recent versions of PHPUnit unconditionally execute tearDown()

+++ tests/Functional/Driver/PgSQL/ResultTest.php
@@ -35,8 +35,10 @@ class ResultTest extends FunctionalTestCase
 
     protected function tearDown(): void
     {
-        $this->connection->executeStatement('DROP TABLE IF EXISTS types_test');
-        $this->connection->executeStatement('DROP TABLE IF EXISTS types_test2');
+        if (TestUtil::isDriverOneOf('pgsql')) {
+            $this->connection->executeStatement('DROP TABLE IF EXISTS types_test');
+            $this->connection->executeStatement('DROP TABLE IF EXISTS types_test2');
+        }
 
         parent::tearDown();
     }
